Move Personal form to PersonalDetails/Edit.vue
After discussing with @ollietreend, we decided that we needed a more
CRUD-like set of views of the forms.  We want to be able to show a read
only view (`Show.vue` most likely), at at least one point in the
application process.  Breaking out the Form templates out in this way
*should* make this simpler to maintain.
